# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bot Factory API — SaaS платформа для автогенерации Telegram-ботов
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface TelegramAuthInput {
  /** Raw Telegram WebApp initData string */
  initData: string;
  /** @nullable */
  referralCode?: string | null;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  basic: 'basic',
  pro: 'pro',
  business: 'business',
} as const;

export interface User {
  id: string;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  credits: number;
  plan: UserPlan;
  /** @nullable */
  planExpiresAt?: string | null;
  referralCode?: string;
  isBanned?: boolean;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface UserStats {
  botsCount: number;
  botsRunning: number;
  totalGenerations: number;
  creditsSpent: number;
}

export type BotStatus = typeof BotStatus[keyof typeof BotStatus];


export const BotStatus = {
  generating: 'generating',
  running: 'running',
  stopped: 'stopped',
  error: 'error',
} as const;

export type BotBotType = typeof BotBotType[keyof typeof BotBotType];


export const BotBotType = {
  simple: 'simple',
  complex: 'complex',
  miniapp: 'miniapp',
} as const;

export interface Bot {
  id: string;
  userId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /**
     * Masked token (last 4 chars)
     * @nullable
     */
  botToken?: string | null;
  /** @nullable */
  botUsername?: string | null;
  status: BotStatus;
  botType: BotBotType;
  /** @nullable */
  generationId?: string | null;
  /** @nullable */
  lastActiveAt?: string | null;
  createdAt: string;
}

export interface BotInput {
  /** @minLength 1 */
  name: string;
  /**
     * User's idea description for the bot
     * @minLength 10
     */
  description: string;
  /**
     * Telegram bot token from @BotFather
     * @nullable
     */
  botToken?: string | null;
}

export interface BotUpdate {
  /** What to add or change in the bot */
  improvement?: string;
}

export interface BotLogs {
  botId: string;
  lines: string[];
}

export type GenerationStepStatus = typeof GenerationStepStatus[keyof typeof GenerationStepStatus];


export const GenerationStepStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
  error: 'error',
  skipped: 'skipped',
} as const;

export interface GenerationStep {
  key: string;
  label: string;
  status: GenerationStepStatus;
}

export type GenerationStatus = typeof GenerationStatus[keyof typeof GenerationStatus];


export const GenerationStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
  failed: 'failed',
} as const;

export interface Generation {
  id: string;
  botId: string;
  status: GenerationStatus;
  steps: GenerationStep[];
  creditsUsed: number;
  /** @nullable */
  elapsedSeconds?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  /** @nullable */
  generationId?: string | null;
  createdAt: string;
}

export interface ChatMessageInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  sessionId?: string | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  purchase: 'purchase',
  generation: 'generation',
  hosting: 'hosting',
  refund: 'refund',
  referral: 'referral',
  bonus: 'bonus',
} as const;

export interface Transaction {
  id: string;
  userId: string;
  /** Positive = income, Negative = expense */
  amount: number;
  type: TransactionType;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export type CreditPurchaseInputPackageId = typeof CreditPurchaseInputPackageId[keyof typeof CreditPurchaseInputPackageId];


export const CreditPurchaseInputPackageId = {
  starter: 'starter',
  basic: 'basic',
  pro: 'pro',
  max: 'max',
} as const;

export interface CreditPurchaseInput {
  packageId: CreditPurchaseInputPackageId;
}

export interface CreditPurchaseResult {
  success: boolean;
  creditsAdded: number;
  newBalance: number;
  transactionId?: string;
}

export type MarketplaceTemplateBotType = typeof MarketplaceTemplateBotType[keyof typeof MarketplaceTemplateBotType];


export const MarketplaceTemplateBotType = {
  simple: 'simple',
  complex: 'complex',
  miniapp: 'miniapp',
} as const;

export interface MarketplaceTemplate {
  id: string;
  name: string;
  description: string;
  botType: MarketplaceTemplateBotType;
  /** Price in credits */
  price: number;
  /** @nullable */
  authorUsername?: string | null;
  deployCount: number;
  /** @nullable */
  rating?: number | null;
  createdAt: string;
}

export interface TemplateDeployInput {
  botToken: string;
}

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  referralsCount: number;
  /** Total credits earned through referrals */
  totalEarned: number;
}

export interface AdminStats {
  totalUsers: number;
  /** Users active in last 30 days */
  activeUsers: number;
  totalBots: number;
  runningBots: number;
  totalGenerations: number;
  /** Total credits purchased */
  totalRevenue: number;
}

export interface AdminCreditsInput {
  userId: string;
  amount: number;
  /** @nullable */
  reason?: string | null;
}