# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bot Factory API — SaaS платформа для автогенерации Telegram-ботов
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/telegram:
    post:
      operationId: authTelegram
      tags: [auth]
      summary: Authenticate via Telegram initData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthInput"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid initData

  # ── USERS ────────────────────────────────────────────────────────────────
...
  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List all user's bots
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: createBot
      tags: [bots]
      summary: Create a new bot (starts generation pipeline)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Bot created and generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "402":
          description: Insufficient credits
  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get bot details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
    patch:
      operationId: updateBot
      tags: [bots]
      summary: Update/improve bot (adds features to existing code)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotUpdate"
      responses:
        "200":
          description: Updated bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete bot and free server resources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /bots/{id}/restart:
    post:
      operationId: restartBot
      tags: [bots]
      summary: Restart bot process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{id}/stop:
    post:
      operationId: stopBot
      tags: [bots]
      summary: Stop bot process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{id}/start:
    post:
      operationId: startBot
      tags: [bots]
      summary: Start a stopped bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{id}/logs:
    get:
      operationId: getBotLogs
      tags: [bots]
      summary: Get last 20 lines of bot logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotLogs"

  # ── CHAT ──────────────────────────────────────────────────────────────────
...
  /generations/{id}:
    get:
      operationId: getGeneration
      tags: [bots]
      summary: Get generation job status and progress steps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generation"

  # ── TRANSACTIONS / CREDITS ────────────────────────────────────────────────
...
  /admin/credits:
    post:
      operationId: adminAddCredits
      tags: [admin]
      summary: Manually add credits to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreditsInput"
      responses:
        "200":
          description: Credits added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
components:
  schemas:
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── AUTH ────────────────────────────────────────────────────────────────
    TelegramAuthInput:
      type: object
      required: [initData]
      properties:
        initData:
          type: string
          description: Raw Telegram WebApp initData string
        referralCode:
          type: ["string", "null"]
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── USER ─────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, telegramId, credits, plan, createdAt]
      properties:
        id:
          type: string
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        credits:
          type: integer
        plan:
          type: string
          enum: [free, basic, pro, business]
        planExpiresAt:
          type: ["string", "null"]
        referralCode:
          type: string
        isBanned:
          type: boolean
        createdAt:
          type: string
...
    Bot:
      type: object
      required: [id, userId, name, status, botType, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        botToken:
          type: ["string", "null"]
          description: Masked token (last 4 chars)
        botUsername:
          type: ["string", "null"]
        status:
          type: string
          enum: [generating, running, stopped, error]
        botType:
          type: string
          enum: [simple, complex, miniapp]
        generationId:
          type: ["string", "null"]
        lastActiveAt:
          type: ["string", "null"]
        createdAt:
          type: string
    BotInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 10
          description: User's idea description for the bot
        botToken:
          type: ["string", "null"]
          description: Telegram bot token from @BotFather
    BotUpdate:
      type: object
      properties:
        improvement:
          type: string
          description: What to add or change in the bot
    BotLogs:
      type: object
      required: [botId, lines]
      properties:
        botId:
          type: string
        lines:
          type: array
          items:
            type: string

    # ── GENERATION ───────────────────────────────────────────────────────────
    GenerationStep:
      type: object
      required: [key, label, status]
      properties:
        key:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [pending, in_progress, done, error, skipped]
    Generation:
      type: object
      required: [id, botId, status, steps, creditsUsed, createdAt]
      properties:
        id:
          type: string
        botId:
          type: string
        status:
          type: string
          enum: [pending, in_progress, done, failed]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/GenerationStep"
        creditsUsed:
          type: integer
        elapsedSeconds:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── CHAT ─────────────────────────────────────────────────────────────────
...
    AdminCreditsInput:
      type: object
      required: [userId, amount]
      properties:
        userId:
          type: string
        amount:
          type: integer
        reason:
          type: ["string", "null"]