# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bot Factory API — SaaS платформа для автогенерации Telegram-ботов
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from './userPlan';

export interface User {
  id: string;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  credits: number;
  plan: UserPlan;
  /** @nullable */
  planExpiresAt?: string | null;
  referralCode?: string;
  isBanned?: boolean;
  createdAt: string;
}